requirements.txt
Pillow~=6.1.0
pytest~=5.0.1
requests~=2.22.0